Create GCC_Veg_Detect_v3.py
This is the python script that uses GCC (green color coordinate) index to detect green pixels. The script will analyze a folder of photos, create a binary mask on green / not green. Then we reduce noise to remove white pixels 1px x 1px. The binary mask for each photo is saved to the same folder with the "modified" suffix.


Then we pull the GPS coordinate for each photo, calculate the proportion of pixels which is green, and export that all to a csv with the photo name. 
There is a simple GUI to choose the folder the photos are in, where to save the csv, and the threshold value you want to use for the GCC.